+++ h2integrate/core/supported_models.py
@@ -207,6 +207,7 @@ no_cost_models = {
     "GasStreamCombinerPerformanceModel",
     "CablePerformanceModel",
     "PipePerformanceModel",
+    "GenericTransporterPerformanceModel",
 }
 
 no_replacement_schedule_models = {